+++ user/src/main/java/com/sparta/rooibos/user/application/service/port/EventProvider.java
@@ -1,4 +1,4 @@
-package com.sparta.rooibos.user.application.service;
+package com.sparta.rooibos.user.application.service.port;
 
 import com.sparta.rooibos.user.application.dto.UserAuthDTO;
 
@@ -6,4 +6,5 @@ public interface EventProvider {
     void sendUserInfo(UserAuthDTO dto);
     void sendUserDeleteInfo(String email);
     void sendUserReportInfo(String email);
+    void blacklistUser(String email, long ttlSeconds);
 }
