+++ user/src/main/java/com/sparta/rooibos/user/infrastructure/redis/UserEventProducer.java
@@ -2,39 +2,33 @@ package com.sparta.rooibos.user.infrastructure.redis;
 
 import com.sparta.rooibos.user.application.dto.UserAuthDTO;
 import com.sparta.rooibos.user.application.dto.UserStreamEvent;
-import com.sparta.rooibos.user.application.service.EventProvider;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.data.redis.connection.stream.ObjectRecord;
-import org.springframework.data.redis.connection.stream.RecordId;
-import org.springframework.data.redis.connection.stream.StreamRecords;
 import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.core.StreamOperations;
+import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 
-@Service
+@Component
 @RequiredArgsConstructor
 @Slf4j
-public class UserEventProducer implements EventProvider {
+public class UserEventProducer  {
     private final @Qualifier("userStreamRedisTemplate") RedisTemplate<String, UserStreamEvent> redisTemplate;
 
     private static final String STREAM_KEY = "user-events";
     private static final long MAX_STREAM_LENGTH = 1000L; // ✅ 유저 이벤트 최대 1000개 유지
 
-    @Override
     public void sendUserInfo(UserAuthDTO dto) {
         UserStreamEvent event = new UserStreamEvent("create", dto);
         addEventToStream(event);
     }
 
-    @Override
     public void sendUserDeleteInfo(String email) {
         UserStreamEvent event = new UserStreamEvent("delete", email);
         addEventToStream(event);
     }
 
-    @Override
     public void sendUserReportInfo(String email) {
         UserStreamEvent event = new UserStreamEvent("report", email);
         addEventToStream(event);
