+++ user/src/main/java/com/sparta/rooibos/user/presentation/controller/UserApproveController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.UUID;
 
 @RestController
-@RequestMapping("/api/v1")
+@RequestMapping("/api/v1/user")
 @RequiredArgsConstructor
 public class UserApproveController {
     private final UserApproveService userApproveService;
